### 2021-07-14 #3494 @whedon
[PRE REVIEW]: A Fast Iterative Method Python package
**Submitting author:** @thomgrand (<a href="http://orcid.org/0000-0002-6787-2570">Thomas Grandits</a>)
**Repository:** <a href="https://github.com/thomgrand/fim-python" target ="_blank">https://github.com/thomgrand/fim-python</a>
**Version:** v1.0
