### 2021-07-15 2e8108e
Fixed missing DOIs in the bibliography and a wrong reference.
