+++ paper.bib
@@ -13,15 +13,18 @@
 	pages = {2468--2488}
 }
 
-@article{evans_1997_partial,
-  title={Partial differential equations and Monge-Kantorovich mass transfer},
-  author={Evans, Lawrence C},
-  journal={Current developments in mathematics},
-  volume={1997},
-  number={1},
-  pages={65--126},
-  year={1997},
-  publisher={International Press of Boston}
+@book{evans_partial_2010,
+    author = {Evans, Lawrence C.},
+     title = {Partial differential equations},
+    series = {Graduate Studies in Mathematics},
+    volume = {19},
+   edition = {Second},
+   publisher = {American Mathematical Society, Providence, RI},
+      year = {2010},
+     pages = {xxii+749},
+      ISBN = {978-0-8218-4974-3},
+       doi = {10.1090/gsm/019},
+       url = {https://doi.org/10.1090/gsm/019},
 }
 
 @article{fu_fast_2013,
@@ -47,7 +50,8 @@
   number={4},
   pages={1591--1595},
   year={1996},
-  publisher={National Acad. Sciences}
+  publisher={National Acad. Sciences},
+  doi = {10.1073/pnas.93.4.1591},
 }
 
 @book{franzone2014mathematical,
@@ -55,7 +59,8 @@
   author={Franzone, Piero Colli and Pavarino, Luca Franco and Scacchi, Simone},
   volume={13},
   year={2014},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-3-319-04801-7}
 }
 
 @article{grandits_inverse_2020,
