### 2021-08-14 #3494 @lucaferranti
Hi @xuanxu , I am willing to review this paper. I have read the JOSS Conflict of Interest policy and I confirm that I have no COIs with this review. 😄
