### 2021-08-18 #3494 @xuanxu
👋 @mgalloy would you be willing to review this paper? I think your knowledge of math, CUDA and Python makes you a good fit for this submission.
