### 2021-09-07 #3641 @xuanxu
> is it necessary to have a separate statement of need section or can it be included in the summary?

Yes, we prefer papers with the *Statement of Need* as a separate section.
