### 2021-09-17 #3641 @xuanxu
@thomgrand: did you have the opportunity to address @lucaferranti's [comments](https://github.com/openjournals/joss-reviews/issues/3641#issuecomment-914074505)?

@RobertRosca: Any update on your review?
