### 2021-09-18 #3641 @thomgrand
Hi @lucaferranti & @xuanxu , many thanks for the initial thorough review!

I agree with most of the points and I'm preparing a commit to address many of the comments so far. It should be ready in the next days.

> ### Paper
>
>     * [ ]  **line 5**: "arrival times $\phi$ of a wave". Should this be "arrival times of wave $\phi$"? $\phi$ is the wave function, isn't it?

In this context, $\phi$ denotes the arrival time, not uncommonly found in the literature:
https://math.stackexchange.com/questions/2816941/relation-between-a-solution-of-eikonal-equation-and-curvature
https://wiki.seg.org/wiki/Dictionary:Eikonal_equation

In the second link you'll find that there is a close connection to the wave function however.
