### 2021-09-21 94f2534
Updates to the README: - Cython is now listed as a prerequisite - Listed the main interface create_fim_solver - The example is now the same as in the documentation - Steps for reproducing the benchmark are now given
Added a markdown document about contributions to the library.
Updates to the documentation:
- The benchmark page now contains instructions to run the benchmark

Revamped the paper according to the JOSS reviews and added a section "Comparison to other tools" with direct links to the libraries.
Minor updates to the benchmark execution.
