### 2021-09-21 #3641 @thomgrand
Many thanks again for the review, I tried to address the points in https://github.com/thomgrand/fim-python/commit/94f2534e29f509ab3da1d01f4f9dd9ba0adf3cbf
>     * [ ]  The installation instructions in the documentation (and in the paper) mention to first run `pip install cython`, this is not mentioned in the installation instructions in the readme.

This should be mentioned now.

>     * [ ]  It would be nice to include the script to reproduce the benchmarks in the documentation. If the script is too long,  the documentation could just have a link to it

The instructions to reproduce the benchmark have been added to the documentation in the section Benchmark.

>     * [ ]  The last two examples in the readme example (GPU jacobi and GPU with active list) cannot be correct, as they are the same code but are supposed to use two different methods
>
>     * [ ]  I find a little confusing that the readme example uses `FIMNP`, `FIMNPAL` etc, but the example in the documentation uses `FIMPY.create_solver`. Moreover, looking at the documentation I got the impression that the functions currently used in the README are meant to be internals, while FIMPY.create_fim_solver is the "external interface" the user is meant to use. This is based to the fact that `FIMPY.create_fim_solver` allows to choose whether to use cpu or gpu and jacobi or active list. Could you comment on this? If this is the case, I would suggest to use the same interface in both.
>
>     * [ ]  Moreover, the README example is not reproducible since it doesn't show how to create `triangs`, `points` etc. Since the example in the documentation is not that long, I would reuse that in the readme too.
>
>     * (This is a smaller one, feel free to ignore it if you disagree) In the example in the documentation, where you have the lines

I agree that this was indeed confusing. The documentation and the ReadMe should now list the same, executable example and also use the same interface as outlined in the documentation.

>
> ```python
> #Create a FIM solver, by default the GPU solver will be called with the active list
> fim = FIMPY.create_fim_solver(points, elems, D)
> ```
>
> I would add an additional comment
>
> ```python
> # set device='cpu' to run on cpu and use_active_list=false to use Jacobi method
> ```
>
> It is true that this is well documented under "Interface methods" section, but I feel highlighting in the documentation example would serve users better, as that's probably the first thing they read.
>
I added said comment.
>     * [ ]  Contribution guidelines are missing. You could either just briefly explain with a few lines in the README how to contribute or setup the github contribution guidelines as described [here](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors). Also from the repository `insights > community` you should find some templates for writing contribution guidelines.

I added a contributing.md with some instructions on how to run the tests and how to report a bug.

>
> ### Paper
>
>     * [x]  **line 5**: "arrival times $\phi$ of a wave". Should this be "arrival times of wave $\phi$"? $\phi$ is the wave function, isn't it?
>
>     * [ ]  **line 10**: this line introduce the Finite Iterative Method (FIM) for the first time, it would be good to have here a reference to the literature where the theory of the method is discussed.
>
The reference to the original paper was added.
>     * [ ]  To me the current "Statement of need" section is more a comparison to other existing tools, I would consider renaming it accordingly. The "statements  of need" checkbox requires
>       > Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
>
>
> For what is worth, I think from your summary section it is already clear enough what problem you are trying to solve and who is the audience. @xuanxu is it necessary to have a separate statement of need section or can it be included in the summary?
>
>     * [ ]  In the comparison to other tools, you mention that there are publicly available libraries which have some of the properties but not all. However, there is no mention of what these publicly available libraries are. I think this section would be much stronger if you listed some of the existing libraries and for each of them what features they have and what features they lack. maybe this could be arranged in a table, this would make much easier to compare your library to existing ones.
>
I reworked the section *Statement of need* and added a new section *Comparison to other tools* that describes the two other known, publicly available libraries for similar problems and their comparison to FIM-Python.

>     * [ ]  I don't think the installation instructions are needed to be mentioned in the paper, at least  not in the statement of need / comparison to other methods section.

The installation instructions have been removed.
