### 2021-09-26 #3641 @RobertRosca
Hello! I've been a bit busy recently, sorry about the delay, here are my notes so far (although most have been addressed already since I last had time to look at this):

As mentioned by @lucaferranti, cython is a build-time requirement and (currently) has to be manually installed.

You can get around this with a `pyproject.toml` file, defining the build system, this will ensure that (if you're using an up to date version of pip) the build time dependencies are all available from the start.

- [ ] Add `pyproject.toml` with build-time dependencies, (see [PEP 508](https://www.python.org/dev/peps/pep-0518/))
- [ ] Add `ipython` to requirements as it is needed for the benchmarks
- [x] No comments on the paper, looks good to me!

Apart from adding `pyproject.toml` for build-time deps and `ipython` being missing, I don't have any other comments right now.

## Misc

Misc notes as I took them (feel free to ignore these :stuck_out_tongue: they're on the "nice to have" side and not important):

- Add a `.gitignore` - when installing in editable mode the `build`/`.egg-info` directories, and the `.cpp`, `.cpython` files get generated and may be accidentally committed to the repository by contributors
- The separate GPU/CPU dependencies, as well as the common dependencies, are described both in the `setup.py` file and in `requirements_*.txt` files, could cause issues if they get out of sync
- Also, `install_requires` gets installed regardless of which extras are selected, so there's no need to do `test_requires_cpu = lib_requires_cpu + ...`
- Ah, those files get used in the workflows for installing before tests, was this done to avoid the issue with requiring Cython/Numpy at build time causing setup to fail? If so you can avoid that with a `pyproject.toml` file specifying build system requirements (see [PEP 508](https://www.python.org/dev/peps/pep-0518/))
- Wheel seems to be required otherwise there's some cupy/numba version issue
- Codecov token is exposed in `bitbucket-pipelines.yml`, not really important though
- Running `pip install ...` shows error: `ERROR: numba 0.54.0 has requirement numpy<1.21,>=1.17, but you'll have numpy 1.21.2 which is incompatible.`. Likely caused by the (relatively) recent addition of the dependency resolver in pip 20.3, you should add a `python3 -m pip install --upgrade pip` step to the CI as without that you may miss issues introduced by pip updates
- GPU testing is a bit of a pain, if you have the computers available at work you can [set up self-hosted runners on them](https://docs.github.com/en/actions/hosting-your-own-runners/adding-self-hosted-runners) and add the runners to the repo, then tag GPU tests/benchmarks as requiring those runners to execute
- Ipython missing from requirements for benchmarks
- Running and benchmarks now
- lucaferranti covered everything else
