### 2021-09-26 #3641 @RobertRosca
I'll double check everything tonight but I think the two installation related changes are the only suggestions I have.
