### 2021-09-28 #3641 @lucaferranti
Hi @thomgrand 👋 , thank you for the changes and apologies for the delay in replying. I am mainly very happy with your changes, however I still have one comment about the paper.

Currently, the statement of need is only one sentence long. It feels a little funny to have a section with only one sentence paragraph. I would suggest you to expand a little and elaborate more.

For example, you could move the third paragraph in the summary ("the eikonal equation has many applications....") to the statement of need section and then add 1-2 sentences about `fim-python` allowing to solve this problem both on CPU and GPU and in more general scenarios than current existing tools.

I think this way the section would be much stronger, of course this is just a suggestion, if you have a better structure in mind, feel free to go with it. However, I would strongly advise against having only a single one-sentence paragraph in whole section.
