### 2021-10-02 17f70f4
Added a bug report template
