+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,26 @@
+---
+name: Bug report
+about: Template for bug reports
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the unexpected behavior
+
+**Please provide the following information:**
+- Operating system
+- Version numbers of
+  - Python
+  - Numpy
+  - Cupy
+
+See [Contributing.md](CONTRIBUTING.md) for a script to automatically output the variables
+
+**Additional information [Optional]**
+Add any other information about the problem here.
