### 2021-10-02 9e3a4fd
Removed the requirements.txt files and merged everything into setup.py The github workflow should reflect this change and now only use the local pip installation method as a test. Added a pyproject.toml file to avoid cython issues when installing over pip. Ipython is now listed as a requirement for testing. Removed the deprecated bitbucket-pipelines.yml file, which was not in use. Restructured the paper, incorporating some of the summary into the statement of need. Slight overhaul of the contribution guidelines.
