+++ CONTRIBUTING.md
@@ -9,13 +9,31 @@ Reporting Bugs
 
 Please submit bug reports to the [issue page](https://github.com/thomgrand/fim-python/issues). Make sure that you include all of the following:
 - Description of the bug
+- Steps to reconstruct the error
 - Operating system
 - Version numbers of
   - Python
   - Numpy
   - Cupy
-- Steps to reconstruct the error
 
+Fetching the version numbers and the operating system info can be automatically achieved by executing the following script in your python environment:
+
+```python
+import platform
+import os
+print("OS Info: %s, %s, v%s" % (os.name, platform.system(), platform.release()))
+
+import numpy
+print("Numpy version: %s" % (numpy.__version__))
+import numba
+print("Numba version: %s" % (numba.__version__))
+
+try:
+    import cupy
+    print("GPU version, version of cupy: %s" % (cupy.__version__))
+except ImportError:
+    print("CPU version only")
+```
 
 Submitting Code
 --------------------
@@ -31,7 +49,7 @@ python -m pytest tests
 
 Before opening a pull request, please make sure that all tests are passing.
 In case you only have the CPU version, all tests for the GPU will be skipped. 
-The gitlab-runner will also test committed versions of the library, but only on the CPU for the lack of a GPU on the runner.
+The github-runner will also test committed versions of the library, but only on the CPU for the lack of a GPU on the runner.
 If you submit new features, please also write tests to ensure functionality of the features.
 
 > **_Note:_**  If you do **not** have a Cupy compatible GPU to test on, please clearly state this in your pull request, so somebody else from the community can test your code with all features enabled.
