+++ requirements_cpu.txt
@@ -1,3 +1,0 @@
-numpy
-numba
-cython
