+++ bitbucket-pipelines.yml
@@ -1,28 +1,0 @@
-#  Template python-build
-
-#  This template allows you to validate your python code.
-#  The workflow allows running tests and code linting on the default branch.
-
-image: python:3.8
-
-pipelines:
-  default:
-    - parallel:
-      - step:
-          name: Test
-          caches:
-            - pip
-          script:
-            - pip install -r requirements_cpu.txt -r requirements_tests.txt
-            - cythonize -i fimpy/fim_cutils.pyx
-            - python tests/generate_test_data.py
-            #- python -m pytest --junitxml=test-reports/report.xml
-            - python -m pytest --cov-config=.coveragerc_cpu --cov=fimpy tests/
-            - bash <(curl -s https://codecov.io/bash) -t 92805a53-76ae-4d04-a9a2-108142ee94d7
-      # Linting disabled
-      #- step:
-      #    name: Lint code
-      #    script:
-      #      # Enforce style consistency across Python projects https://flake8.pycqa.org/en/latest/manpage.html
-      #      - pip install flake8
-      #      - flake8 . --extend-exclude=dist,build --show-source --statistics
