+++ pyproject.toml
@@ -1,0 +1,3 @@
+[build-system]
+requires = ["setuptools", "wheel", "Cython>=0.29.22"]
+build-backend = "setuptools.build_meta"

[…]
