### 2021-10-02 3036d83
Added the missing .gitignore file
