### 2021-10-02 #3641 @thomgrand
Hi @RobertRosca & @lucaferranti, many thanks for the review. Please find my response in-between the quotes. I tried to address most of the points in https://github.com/thomgrand/fim-python/commit/9e3a4fdb58450e1fc278f84a6f9ec4c0b738922c

# Robert Rosca

>
> As mentioned by @lucaferranti, cython is a build-time requirement and (currently) has to be manually installed.
>
> You can get around this with a `pyproject.toml` file, defining the build system, this will ensure that (if you're using an up to date version of pip) the build time dependencies are all available from the start.

Thanks, I didn't actually know about this. I added the `pyproject.toml` with the setuptools, wheel and cython requirements.

>     * [ ]  Add `pyproject.toml` with build-time dependencies, (see [PEP 508](https://www.python.org/dev/peps/pep-0518/))
>
>     * [ ]  Add `ipython` to requirements as it is needed for the benchmarks
>

I also added ipython as an requirement in `setup.py` for the tests and docs extra.

> ## Misc
>
> Misc notes as I took them (feel free to ignore these 😛 they're on the "nice to have" side and not important):
>
>     * Add a `.gitignore` - when installing in editable mode the `build`/`.egg-info` directories, and the `.cpp`, `.cpython` files get generated and may be accidentally committed to the repository by contributors

I added a `.gitignore` to ignore the temporary files.

>     * The separate GPU/CPU dependencies, as well as the common dependencies, are described both in the `setup.py` file and in `requirements_*.txt` files, could cause issues if they get out of sync
>
>     * Also, `install_requires` gets installed regardless of which extras are selected, so there's no need to do `test_requires_cpu = lib_requires_cpu + ...`
>
>     * Ah, those files get used in the workflows for installing before tests, was this done to avoid the issue with requiring Cython/Numpy at build time causing setup to fail? If so you can avoid that with a `pyproject.toml` file specifying build system requirements (see [PEP 508](https://www.python.org/dev/peps/pep-0518/))
>

I agree this was not a nice solution. I removed the requirements.txt files and tried to keep everything in `setup.py` now. The github workflows should now reflect this change and only install the library via `pip`.

>     * Wheel seems to be required otherwise there's some cupy/numba version issue

I couldn't reproduce this issue, but I hope that adding wheel to `pyproject.toml` resolves this issue.

>     * Codecov token is exposed in `bitbucket-pipelines.yml`, not really important though

This was an old unused token that is now removed together with the pipeline.

>     * Running `pip install ...` shows error: `ERROR: numba 0.54.0 has requirement numpy<1.21,>=1.17, but you'll have numpy 1.21.2 which is incompatible.`. Likely caused by the (relatively) recent addition of the dependency resolver in pip 20.3, you should add a `python3 -m pip install --upgrade pip` step to the CI as without that you may miss issues introduced by pip updates

I added the maximum numpy version and also made sure that pip is upgraded before installing the library on the github runners.

>     * GPU testing is a bit of a pain, if you have the computers available at work you can [set up self-hosted runners on them](https://docs.github.com/en/actions/hosting-your-own-runners/adding-self-hosted-runners) and add the runners to the repo, then tag GPU tests/benchmarks as requiring those runners to execute

Unfortunately, I currently don't have an idle machine with a GPU to act as a github runner, but I hope to do so at one point.

# Lucaferranti

>
>
> Hi @thomgrand 👋 , thank you for the changes and apologies for the delay in replying. I am mainly very happy with your changes, however I still have one comment about the paper.
>
> Currently, the statement of need is only one sentence long. It feels a little funny to have a section with only one sentence paragraph. I would suggest you to expand a little and elaborate more.
>
> For example, you could move the third paragraph in the summary ("the eikonal equation has many applications....") to the statement of need section and then add 1-2 sentences about `fim-python` allowing to solve this problem both on CPU and GPU and in more general scenarios than current existing tools.
>
> I think this way the section would be much stronger, of course this is just a suggestion, if you have a better structure in mind, feel free to go with it. However, I would strongly advise against having only a single one-sentence paragraph in the whole section.

The single-sentence section was an oversight on my side. I like your suggestion and moved the mentioned paragraph into the statement of need section and added a sentence that the tool will hopefully prove to be an accessible tool.
