### 2021-10-07 #3641 @xuanxu
@lucaferranti @RobertRosca any update on your reviews after [latest changes/comments](https://github.com/openjournals/joss-reviews/issues/3641#issuecomment-932748322) from @thomgrand?
