### 2021-10-22 8f926ad
Bumped the version number to 1.1. Minor modifications to the contribution guidelines.
