+++ CONTRIBUTING.md
@@ -3,7 +3,6 @@
 Thank you for your interest in FIM-Python. Any help and contributions are appreciated.
 
 
-
 Reporting Bugs
 ---------------------
 
@@ -37,7 +36,7 @@ except ImportError:
 
 Submitting Code
 --------------------
-FIM-Python uses [pytest](https://docs.pytest.org) to test the code. Pip can take care of installing all necessary packages by listing the extra ``tests``:
+FIM-Python uses the [pytest](https://docs.pytest.org) framework. Pip can take care of installing all necessary packages by listing the extra ``tests``:
 ```bash
 pip install fim-python[gpu,tests]
 ```
@@ -47,9 +46,9 @@ python tests/generate_test_data.py #First time only to generate the test example
 python -m pytest tests
 ```
 
-Before opening a pull request, please make sure that all tests are passing.
+Before opening a pull request for newly written code, please make sure that all tests are passing.
 In case you only have the CPU version, all tests for the GPU will be skipped. 
-The github-runner will also test committed versions of the library, but only on the CPU for the lack of a GPU on the runner.
-If you submit new features, please also write tests to ensure functionality of the features.
+If you submit new features, please also write tests to ensure functionality of these features.  
+The github-runner will also test pull-requests and committed versions of the library, but only on the CPU for the lack of a GPU on the runner.
 
 > **_Note:_**  If you do **not** have a Cupy compatible GPU to test on, please clearly state this in your pull request, so somebody else from the community can test your code with all features enabled.
