### 2022-01-26 60b9037
Removed the numba dependency of the package and replaced the fast 2D and 3D methods with Cython.
