+++ fimpy/utils/comp.py
@@ -2,92 +2,44 @@
 Note that all custom, efficient implementations assume that :math:`A` is a symmetric matrix.
 """
 import numpy as np
-from numba import njit, prange, vectorize, guvectorize
-import numba
+from .cython.comp import metric_sqr_norm_matrix_2D_vec, metric_sqr_norm_matrix_3D_vec
 try:
   import cupy as cp
   cupy_enabled = True
 except ImportError as err:
   cupy_enabled = False
 
-@njit(cache=True, nogil=True) #, parallel=True)
-def metric_norm_matrix_2D_njit(A, x1, x2, ret_sqrt=True):
-  """Custom implementation of :func:`metric_norm_matrix` for ``lib = np`` and :math:`d = 2`.
-  """
-  a, b, c = A[..., 0, 0], A[..., 1, 1], A[..., 0, 1]
-  norm = (x1[..., 0] * (a * x2[..., 0] + c * x2[..., 1]) + x1[..., 1] * (c * x2[..., 0] + b * x2[..., 1]))
+
+def _broadcast_metric_params(A, x1, x2):
+  if A.ndim == 4: #Broadcasting
+    #To remove the read-only problem, copy back
+    A = np.broadcast_to(A, [A.shape[0], max(A.shape[1], x1.shape[1]), x1.shape[-1], x1.shape[-1]]).copy()
+    x1 = np.broadcast_to(x1, A.shape[:-1]).copy()
+    x2 = np.broadcast_to(x2, A.shape[:-1]).copy()
+
+  return A, x1, x2
+
+
+def metric_norm_matrix_2D_cython(A, x1, x2, ret_sqrt=True):
+  A, x1, x2 = _broadcast_metric_params(A, x1, x2)
+  A_flat, x1_flat, x2_flat = A.reshape([-1, 2, 2]), x1.reshape([-1, 2]), x2.reshape([-1, 2])
+  norm = metric_sqr_norm_matrix_2D_vec(A_flat, x1_flat, x2_flat)
 
   if ret_sqrt:
     norm = np.sqrt(norm)
 
-  return norm
+  return norm.reshape(A.shape[:-2])
 
-@vectorize([numba.float32(*(numba.float32,)*12),
-            numba.float64(*(numba.float64,)*12)], cache=True, target='cpu')
-def metric_sqr_norm_matrix_3D_vec(a, b, c, d, e, f, x11, x12, x13, x21, x22, x23):
-  return (x11 * (a * x21 + d * x22 + e * x23)
-          + x12 * (d * x21 + b * x22 + f * x23)
-          + x13 * (e * x21 + f * x22 + c * x23))
+def metric_norm_matrix_3D_cython(A, x1, x2, ret_sqrt=True):
+  A, x1, x2 = _broadcast_metric_params(A, x1, x2)
+  A_flat, x1_flat, x2_flat = A.reshape([-1, 3, 3]), x1.reshape([-1, 3]), x2.reshape([-1, 3])
+  norm = metric_sqr_norm_matrix_3D_vec(A_flat, x1_flat, x2_flat)
 
-@vectorize([numba.float32(*(numba.float32,)*12),
-            numba.float64(*(numba.float64,)*12)], cache=True, target='cpu')
-def metric_norm_matrix_3D_vec(a, b, c, d, e, f, x11, x12, x13, x21, x22, x23):
-  return np.sqrt(x11 * (a * x21 + d * x22 + e * x23)
-          + x12 * (d * x21 + b * x22 + f * x23)
-          + x13 * (e * x21 + f * x22 + c * x23))
-
-@njit(cache=True, nogil=True, parallel=True)
-def metric_norm_matrix_3D_njit_flat(A, x1, x2, ret_sqrt=True):
[…]
