+++ fimpy/utils/cython/__init__.py
@@ -1,0 +1,1 @@
+from .comp import metric_sqr_norm_matrix_2D_vec, metric_sqr_norm_matrix_3D_vec #, sqrt_cython
