+++ fimpy/__init__.py
@@ -1,6 +1,6 @@
 #TODO: Import all symbols and methods here
 from .solver import FIMPY
 
-__version__ = "1.0.1"
+__version__ = "1.2"
 __author__ = "Thomas Grandits"
 #__all__ = ["FIMPY"]

[…]
