### 2022-01-27 6f70e83
Updated the Cython function to take the resultview as a parameter Bump to version 1.2 Added a few docstrings
