+++ fimpy/utils/cython/comp.pyx
@@ -13,29 +13,26 @@ ctypedef fused mat_t:
     float
     double
 
-  
-ctypedef fused ind_t:
-    int
-    long long
+
+ctypedef fused const_mat_t:
+  const float
+  const double
+
 
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
-def metric_sqr_norm_matrix_2D_vec(A_np, mat_t[:, :] x1, mat_t[:, :] x2):
-  result = np.empty(shape=A_np.shape[0], dtype=A_np.dtype)
-  cdef mat_t[:, :, :] A = A_np
+def metric_sqr_norm_matrix_2D_vec(const_mat_t[:, :, :] A, const_mat_t[:, :] x1, const_mat_t[:, :] x2, mat_t[::1] result):
   a, b, c = A[..., 0, 0], A[..., 1, 1], A[..., 0, 1]
 
-  
-  cdef mat_t[::1] result_view = result
-  cdef mat_t[:] a_ = a
-  cdef mat_t[:] b_ = b
-  cdef mat_t[:] c_ = c
+  cdef const_mat_t[:] a_ = a
+  cdef const_mat_t[:] b_ = b
+  cdef const_mat_t[:] c_ = c
   cdef int rows, row_i
 
   rows = A.shape[0]
   for row_i in prange(rows, nogil=True):
-    result_view[row_i] = (x1[row_i, 0] * (a_[row_i] * x2[row_i, 0] + c_[row_i] * x2[row_i, 1]) 
+    result[row_i] = (x1[row_i, 0] * (a_[row_i] * x2[row_i, 0] + c_[row_i] * x2[row_i, 1]) 
                             + x1[row_i, 1] * (c_[row_i] * x2[row_i, 0] + b_[row_i] * x2[row_i, 1]))
 
 
@@ -43,30 +40,27 @@ def metric_sqr_norm_matrix_2D_vec(A_np, mat_t[:, :] x1, mat_t[:, :] x2):
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
-def metric_sqr_norm_matrix_3D_vec(A_np, mat_t[:, :] x1, mat_t[:, :] x2):
-  result = np.empty(shape=A_np.shape[0], dtype=A_np.dtype)
-  cdef mat_t[:, :, :] A = A_np
+def metric_sqr_norm_matrix_3D_vec(const_mat_t[:, :, :] A, const_mat_t[:, :] x1, const_mat_t[:, :] x2, mat_t[::1] result):
   a, b, c, d, e, f = A[..., 0, 0], A[..., 1, 1], A[..., 2, 2], A[..., 0, 1], A[..., 0, 2], A[..., 1, 2]
 
-  cdef mat_t[::1] result_view = result
   cdef int rows, row_i
-  cdef mat_t[:] a_ = a
-  cdef mat_t[:] b_ = b
-  cdef mat_t[:] c_ = c
-  cdef mat_t[:] d_ = d
-  cdef mat_t[:] e_ = e
-  cdef mat_t[:] f_ = f
+  cdef const_mat_t[:] a_ = a
+  cdef const_mat_t[:] b_ = b
+  cdef const_mat_t[:] c_ = c
+  cdef const_mat_t[:] d_ = d
+  cdef const_mat_t[:] e_ = e
+  cdef const_mat_t[:] f_ = f
 
-  cdef mat_t[:] x11_ = x1[..., 0]
-  cdef mat_t[:] x12_ = x1[..., 1]
-  cdef mat_t[:] x13_ = x1[..., 2]
-  cdef mat_t[:] x21_ = x2[..., 0]
-  cdef mat_t[:] x22_ = x2[..., 1]
[…]
