+++ fimpy/utils/cython/__init__.py
@@ -1,1 +1,4 @@
+"""This subpackage contains the Cython implementations for fast 2D and 3D metric norm computations.
+"""
+
 from .comp import metric_sqr_norm_matrix_2D_vec, metric_sqr_norm_matrix_3D_vec #, sqrt_cython
