+++ fimpy/__init__.py
@@ -3,4 +3,3 @@ from .solver import FIMPY
 
 __version__ = "1.2"
 __author__ = "Thomas Grandits"
-#__all__ = ["FIMPY"]
