### 2022-01-27 30720e7
Increased the cython tolerance to avoid CI failures due to precision errors
